Organize project structure:
add shared service layer for authenticated user actions and prepare role-based API routing
