+++ foodDeliveryCleanArch/foodDelivery.Application/Interface/IAuthService.cs
@@ -6,5 +6,6 @@ public interface IAuthService
 {
     public string CreateToken(User user);
     public Token CheckToken(Role role);
+    public Token CheckToken();
 
 }
