+++ foodDeliveryCleanArch/foodDelivery.Application/Interface/ICustomerService.cs
@@ -5,5 +5,4 @@ namespace foodDelivery.Application.Interface;
 public interface ICustomerService
 {
     public CustomerSignupResponse Signup(CustomerSignupRequest request);
-    public AutocompleteResponseDto AutocompleteRestaurants(string prefix);
 }
