+++ foodDeliveryCleanArch/foodDelivery.Application/Interface/IRestaurantService.cs
@@ -12,7 +12,5 @@ public interface IRestaurantService
     public AddFoodResponse AddFood(AddFoodRequest request);
     public void DeleteMenu(Guid menuId, Guid restaurantId);
     public void DeleteFood(Guid foodId, Guid restaurantId);
-    public MenuDetailsDto GetMenu(Guid restaurantId, Guid menuId);
     public void SetFoodStock(Guid restaurantId, Guid foodId, UpdateStockDto request);
-    public MenusDto GetMenus(Guid restaurantId);
 }
