+++ foodDeliveryCleanArch/foodDelivery.Pres/Controller/CustomerController.cs
@@ -26,23 +26,4 @@ public class CustomerController(ICustomerService customerService) : ControllerBa
             return StatusCode(500, $"An unexpected error occurred\\the following error: {e.Message}");
         }
     }
-
-    [HttpGet("autocomplete/restaurant")]
-    [Authorize]
-    public IActionResult AutocompleteRestaurants(string prefix)
-    {
-        try
-        {
-            var response = customerService.AutocompleteRestaurants(prefix);
-            return Ok(response);
-        }
-        catch (ArgumentException e)
-        {
-            return BadRequest(new { error = e.Message });
-        }
-        catch (Exception e)
-        {
-            return StatusCode(500, $"An unexpected error occurred\\the following error: {e.Message}");
-        }
-    }
 }
